Moved afk setting to user command
